moved bindings to py_streamers.cpp
