+++ src/python/py_openvino_genai.cpp
@@ -32,6 +32,7 @@ using ov::genai::get_version;
 void init_lora_adapter(py::module_& m);
 void init_perf_metrics(py::module_& m);
 void init_tokenizer(py::module_& m);
+void init_streamers(py::module_& m);
 void init_generation_config(py::module_& m);
 
 void init_continuous_batching_pipeline(py::module_& m);
@@ -64,40 +65,6 @@ auto encoded_results_docstring = R"(
     scores: sum of logarithmic probabilities of all tokens in the sequence.
     metrics: performance metrics with tpot, ttft, etc. of type ov::genai::PerfMetrics.
 )";
-
-auto streamer_base_docstring =  R"(
-    Base class for streamers. In order to use inherit from from this class and implement put, and methods.
-)";
-
-auto text_streamer_docstring =  R"(
-TextStreamer is used to decode tokens into text and call a user-defined callback function.
-
-tokenizer: Tokenizer object to decode tokens into text.
-callback: User-defined callback function to process the decoded text, callback should return either boolean flag or StreamingStatus.
-
-)";
-
-class ConstructableStreamer: public StreamerBase {
-    bool put(int64_t token) override {
-        PYBIND11_OVERRIDE(
-            bool,  // Return type
-            StreamerBase,  // Parent class
-            put,  // Name of function in C++ (must match Python name)
-            token  // Argument(s)
-        );
-    }
-    StreamingStatus write(int64_t token) override {
-        PYBIND11_OVERRIDE(
-            StreamingStatus,  // Return type
-            StreamerBase,  // Parent class
-            write,  // Name of function in C++ (must match Python name)
-            token  // Argument(s)
-        );
-    }
-    void end() override {
-        PYBIND11_OVERRIDE_PURE(void, StreamerBase, end);
-    }
-};
 
 } // namespace
 
@@ -133,27 +100,9 @@ PYBIND11_MODULE(py_openvino_genai, m) {
         .def_readonly("tokens", &EncodedResults::tokens)
         .def_readonly("scores", &EncodedResults::scores)
         .def_readonly("perf_metrics", &EncodedResults::perf_metrics);
-
-    py::enum_<ov::genai::StreamingStatus>(m, "StreamingStatus")
-        .value("RUNNING", ov::genai::StreamingStatus::RUNNING)
-        .value("CANCEL", ov::genai::StreamingStatus::CANCEL)
-        .value("STOP", ov::genai::StreamingStatus::STOP);
-
-    auto streamer = py::class_<StreamerBase, ConstructableStreamer, std::shared_ptr<StreamerBase>>(m, "StreamerBase", streamer_base_docstring)  // Change the holder form unique_ptr to shared_ptr
-        .def(py::init<>())
-        .def("write", &StreamerBase::write, "Write is called every time new token is decoded. Returns a StreamingStatus flag to indicate whether generation should be stopped or cancelled", py::arg("token"))
-        .def("end", &StreamerBase::end, "End is called at the end of generation. It can be used to flush cache if your own streamer has one");
     
-    py::class_<TextStreamer, std::shared_ptr<TextStreamer>>(m, "TextStreamer", text_streamer_docstring)
-        .def(py::init<const Tokenizer&, std::function<CallbackTypeVariant(std::string)>>(), py::arg("tokenizer"), py::arg("callback"))
-        .def("write", &TextStreamer::write)
-        .def("end", &TextStreamer::end);
-
-    OPENVINO_SUPPRESS_DEPRECATED_START
-    streamer.def("put", &StreamerBase::put, "Put is called every time new token is decoded. Returns a bool flag to indicate whether generation should be stopped, if return true generation stops", py::arg("token"));
-    OPENVINO_SUPPRESS_DEPRECATED_END
-
     init_tokenizer(m);
+    init_streamers(m);
     init_lora_adapter(m);
     init_generation_config(m);
